app/requirements.txt
requests
beautifulsoup4
deepl
python-dotenv
deep-translator